share the CommonTokenStream instance to avoid repetitive buffer initialization

+++ xerial-core/src/test/java/org/xerial/silk/SilkParserPerformanceTest.java
@@ -41,8 +41,8 @@ import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
 import org.xerial.lens.tree.TreeEvent;
-import org.xerial.lens.tree.TreeEventHandlerBase;
 import org.xerial.lens.tree.TreeEvent.EventType;
+import org.xerial.lens.tree.TreeEventHandlerBase;
 import org.xerial.util.FileResource;
 import org.xerial.util.StopWatch;
 import org.xerial.util.log.Logger;
@@ -54,11 +54,11 @@ import org.xml.sax.SAXException;
 import org.xml.sax.XMLReader;
 import org.xmlpull.v1.XmlPullParser;
 
-public class SilkParserPerformanceTest
-{
+public class SilkParserPerformanceTest {
     private static Logger _logger = Logger.getLogger(SilkParserPerformanceTest.class);
 
-    private static final URL largeFile = FileResource.find(SilkParserPerformanceTest.class, "scaffold1.silk");
+    private static final URL largeFile = FileResource.find(SilkParserPerformanceTest.class,
+            "scaffold1.silk");
     private static final int largeFileSize = 30593075;
     private static final double largeFileLines = 111965;
     private static final int numNodes = 5826313;
@@ -66,21 +66,18 @@ public class SilkParserPerformanceTest
     private SilkParserConfig config;
 
     @Before
-    public void setUp() throws Exception
-    {
+    public void setUp() throws Exception {
         config = new SilkParserConfig();
         config.bufferSize = 1024 * 1024 * 8; // 8MB
         config.numWorkers = 2;
     }
 
     @After
-    public void tearDown() throws Exception
-    {}
+    public void tearDown() throws Exception {}
 
     @Ignore
     @Test
-    public void xmlReadTest() throws Exception
-    {
+    public void xmlReadTest() throws Exception {
         SAXParserFactory spf = SAXParserFactory.newInstance();
 
         // Set namespaceAware to true to get a parser that corresponds to
@@ -100,58 +97,48 @@ public class SilkParserPerformanceTest
         // Set the ContentHandler of the XMLReader
         xmlReader.setContentHandler(new ContentHandler() {
 
-            public void characters(char[] ch, int start, int length) throws SAXException
-            {
+            public void characters(char[] ch, int start, int length) throws SAXException {
 
             }
 
-            public void endDocument() throws SAXException
-            {
+            public void endDocument() throws SAXException {
 
             }
 
-            public void endElement(String uri, String localName, String name) throws SAXException
-            {
+            public void endElement(String uri, String localName, String name) throws SAXException {
 
             }
 
-            public void endPrefixMapping(String prefix) throws SAXException
-            {
+            public void endPrefixMapping(String prefix) throws SAXException {
 
             }
 
-            public void ignorableWhitespace(char[] ch, int start, int length) throws SAXException
-            {
+            public void ignorableWhitespace(char[] ch, int start, int length) throws SAXException {
 
             }
 
-            public void processingInstruction(String target, String data) throws SAXException
-            {
+            public void processingInstruction(String target, String data) throws SAXException {
 
             }
 
-            public void setDocumentLocator(Locator locator)
-            {
+            public void setDocumentLocator(Locator locator) {
 
             }
 
-            public void skippedEntity(String name) throws SAXException
-            {
+            public void skippedEntity(String name) throws SAXException {
 
             }
 
-            public void startDocument() throws SAXException
-            {
+            public void startDocument() throws SAXException {
 
             }
 
-            public void startElement(String uri, String localName, String name, Attributes atts) throws SAXException
-            {
+            public void startElement(String uri, String localName, String name, Attributes atts)
+                    throws SAXException {
 
             }
 
-            public void startPrefixMapping(String prefix, String uri) throws SAXException
-            {
+            public void startPrefixMapping(String prefix, String uri) throws SAXException {
 
             }
         });
@@ -169,60 +156,54 @@ public class SilkParserPerformanceTest
 
     @Ignore
     @Test
-    public void xmlPullPerformance() throws Exception
-    {
+    public void xmlPullPerformance() throws Exception {
         XmlPullParser parser = PullParserUtil.newParser(new BufferedReader(new FileReader(
                 "f:/cygwin/home/leo/work/t2k/xmark-1.0.xml")));
 
         StopWatch timer = new StopWatch();
-        while (parser.next() != XmlPullParser.END_DOCUMENT)
-        {}
+        while (parser.next() != XmlPullParser.END_DOCUMENT) {}
         _logger.info("time: " + timer.getElapsedTime());
 
         // 125.3 sec. (Xeon)
 
     }
 
-    public static void reportNodeCountStat(int count, double time)
-    {
+    public static void reportNodeCountStat(int count, double time) {
         double percentage = ((double) count / numNodes) * 100;
         double speed = numNodes / time;
 
-        _logger.debug(String.format("%2.2f%%, time=%5.2f, %,10.0f nodes/s", percentage, time, speed));
+        _logger.debug(String
+                .format("%2.2f%%, time=%5.2f, %,10.0f nodes/s", percentage, time, speed));
 
     }
 
-    public static void reportTotalSpeed(String name, double time)
-    {
+    public static void reportTotalSpeed(String name, double time) {
         reportTotalSpeed(name, numNodes, time);
     }
 
-    public static void reportTotalSpeed(String name, int count, double time)
-    {
-        double speedPerNode = ((double) count) / time;
-        double speedPerLine = ((double) largeFileLines) / time;
+    public static void reportTotalSpeed(String name, int count, double time) {
+        double speedPerNode = (count) / time;
+        double speedPerLine = (largeFileLines) / time;
         double speedInMBS = largeFileSize / 1024 / 1024 / time;
-        _logger.info(String.format("[%-20s] time=%5.2f %,10.0f nodes/s, %5.2f MB/s %7.0f lines/s", name, time,
-                speedPerNode, speedInMBS, speedPerLine));
+        _logger.info(String.format("[%-20s] time=%5.2f %,10.0f nodes/s, %5.2f MB/s %7.0f lines/s",
+                name, time, speedPerNode, speedInMBS, speedPerLine));
     }
 
     @Test
-    public void maxReadSpeed() throws Exception
-    {
-        BufferedReader reader = new BufferedReader(new InputStreamReader(largeFile.openStream()), config.bufferSize);
+    public void maxReadSpeed() throws Exception {
+        BufferedReader reader = new BufferedReader(new InputStreamReader(largeFile.openStream()),
+                config.bufferSize);
         int lineCount = 0;
         String line = null;
         StopWatch timer = new StopWatch();
-        while ((line = reader.readLine()) != null)
-        {
+        while ((line = reader.readLine()) != null) {
             lineCount++;
-            if (lineCount % 100000 == 0)
-            {
+            if (lineCount % 100000 == 0) {
                 double percentage = (lineCount / largeFileLines) * 100;
                 double time = timer.getElapsedTime();
                 double speed = lineCount / time;
-                _logger.debug(String.format("%2.2f%%, line=%d, time=%s, %2.2f lines/s", percentage, lineCount, time,
-                        speed));
+                _logger.debug(String.format("%2.2f%%, line=%d, time=%s, %2.2f lines/s", percentage,
+                        lineCount, time, speed));
             }
         }
         reportTotalSpeed("BufferedReader", timer.getElapsedTime());
@@ -233,8 +214,7 @@ public class SilkParserPerformanceTest
     }
 
     @Test
-    public void maxReadSpeedWithoutParsingNewLine() throws Exception
-    {
+    public void maxReadSpeedWithoutParsingNewLine() throws Exception {
         Reader reader = new InputStreamReader(largeFile.openStream());
         int lineCount = 0;
 
@@ -242,8 +222,7 @@ public class SilkParserPerformanceTest
         char[] buf = new char[config.bufferSize];
         int numBytes = 0;
         int numReadBytes = 0;
-        while ((numReadBytes = reader.read(buf, 0, config.bufferSize)) != -1)
-        {
+        while ((numReadBytes = reader.read(buf, 0, config.bufferSize)) != -1) {
             numBytes += numReadBytes;
         }
         reportTotalSpeed("Reader", timer.getElapsedTime());
@@ -255,8 +234,7 @@ public class SilkParserPerformanceTest
 
     @Ignore
     @Test
-    public void maxReadSpeedTest3() throws Exception
-    {
+    public void maxReadSpeedTest3() throws Exception {
         int unit = 1024 * 1024;
 
         FileInputStream fin = new FileInputStream("scaffold1.txt");
@@ -267,11 +245,9 @@ public class SilkParserPerformanceTest
         ByteBuffer buf = ByteBuffer.allocate(unit);
         StopWatch timer = new StopWatch();
         int pos = 0;
-        while (true)
-        {
+        while (true) {
             buf.rewind();
-            if (channel.read(buf) <= -1)
-            {
+            if (channel.read(buf) <= -1) {
                 break;
             }
         }
@@ -283,16 +259,13 @@ public class SilkParserPerformanceTest
 
     @Ignore
     @Test
-    public void silkStreamReaderPerformance() throws Exception
-    {
+    public void silkStreamReaderPerformance() throws Exception {
         SilkPullParser reader = new SilkPullParser(largeFile);
         StopWatch timer = new StopWatch();
         int count = 0;
         TreeEvent e;
-        while ((e = reader.next()) != null)
-        {
-            if (e.event == EventType.VISIT)
-            {
+        while ((e = reader.next()) != null) {
+            if (e.event == EventType.VISIT) {
                 count++;
                 if (count % 1000000 == 0)
                     reportNodeCountStat(count, timer.getElapsedTime());
@@ -312,14 +285,12 @@ public class SilkParserPerformanceTest
     }
 
     @Test
-    public void pullParserPerformanceTest() throws Exception
-    {
+    public void pullParserPerformanceTest() throws Exception {
         SilkLinePullParser reader = new SilkLinePullParser(largeFile);
         StopWatch timer = new StopWatch();
         int count = 0;
         SilkEvent e;
-        while ((e = reader.next()) != null)
-        {
+        while ((e = reader.next()) != null) {
 
         }
         reportTotalSpeed("SilkLinePullParser", timer.getElapsedTime());
@@ -335,20 +306,17 @@ public class SilkParserPerformanceTest
     }
 
     @Test
-    public void pushParserPerformance() throws Exception
-    {
+    public void pushParserPerformance() throws Exception {
         final SilkLinePushParser reader = new SilkLinePushParser(largeFile, config);
         final StopWatch timer = new StopWatch();
 
         reader.parse(new SilkEventHandler() {
             int lineCount = 0;
 
-            public void handle(SilkEvent event)
-            {
+            public void handle(SilkEvent event) {
                 lineCount++;
 
-                if (event.getType() == SilkEventType.END_OF_FILE)
-                {
+                if (event.getType() == SilkEventType.END_OF_FILE) {
                     _logger.debug("line count = " + lineCount);
                 }
 
@@ -358,23 +326,23 @@ public class SilkParserPerformanceTest
 
         // 20000 lines/sec (Xeon 3.0 * dual) 
         // 45000 lines/sec (Xeon 3.0 * dual when using recursive descent parser)
+
+        // time= 1.63  3,569,374 nodes/s, 17.77 MB/s   68593 lines/s (VAIO Z)
+
     }
 
     @Test
-    public void fastPushParserPerformance() throws Exception
-    {
+    public void fastPushParserPerformance() throws Exception {
         SilkLineFastParser parser = new SilkLineFastParser(largeFile, config);
         StopWatch timer = new StopWatch();
         parser.parse(new SilkEventHandler() {
 
             long lineCount = 0;
 
-            public void handle(SilkEvent event) throws Exception
-            {
+            public void handle(SilkEvent event) throws Exception {
                 lineCount++;
 
-                if (event.getType() == SilkEventType.END_OF_FILE)
-                {
+                if (event.getType() == SilkEventType.END_OF_FILE) {
                     _logger.debug("line count = " + lineCount);
                 }
             }
@@ -386,8 +354,7 @@ public class SilkParserPerformanceTest
     int count = 0;
 
     @Test
-    public void parserPerformance() throws Exception
-    {
+    public void parserPerformance() throws Exception {
         SilkParser parser = new SilkParser(largeFile, config);
         final StopWatch timer = new StopWatch();
 
@@ -395,8 +362,7 @@ public class SilkParserPerformanceTest
         parser.parse(new TreeEventHandlerBase() {
 
             @Override
-            public void visitNode(String nodeName, String immediateNodeValue) throws Exception
-            {
+            public void visitNode(String nodeName, String immediateNodeValue) throws Exception {
                 count++;
                 if (count % 1000000 == 0)
                     reportNodeCountStat(count, timer.getElapsedTime());
@@ -409,17 +375,14 @@ public class SilkParserPerformanceTest
 
     @Ignore
     @Test
-    public void pullParserPerformance() throws Exception
-    {
+    public void pullParserPerformance() throws Exception {
         SilkPullParser parser = new SilkPullParser(largeFile, config);
         final StopWatch timer = new StopWatch();
 
         int count = 0;
         TreeEvent e = null;
-        while ((e = parser.next()) != null)
-        {
-            if (e.event == EventType.VISIT)
-            {
+        while ((e = parser.next()) != null) {
+            if (e.event == EventType.VISIT) {
                 count++;
                 if (count % 1000000 == 0)
                     reportNodeCountStat(count, timer.getElapsedTime());
